+++ qjazz-server/qjazz-map/src/monitor.rs
@@ -2,7 +2,6 @@
 //! Implement monitoring for OWS requests
 //!
 
-
 #[cfg(feature = "monitor")]
 mod mon {
     use actix_web::{